renaming the "cleaners"
